Initial Commit
Initial commit of the work I've been doing on my asteroids game project.

About to update the main systems to use a server holder/provider instead
of 15 goddamn singletons. Mainly because I Want to be able to have the
abstract base class of the renderer without giving away details.
More to come...