Latest code update
Added UI in the top right, centre and left for each of the main elements
in the game. FW1FontWrapper makes DX11 and D2D place nice together
really well, glad I found it and didn't go with a bigger library.
Just needs a little more work and it's a game!